+++ tests/test_project/code_backup/project/main.py
@@ -1,0 +1,3 @@
+from machinable import execute
+
+execute("dummy", storage="../../../_test_data/code_backup/symlinks")